# Add cache for coveringTiles