@JannikGM 2021-09-03
I was asked to comment on the state of my old PRs.

I think this PR should be postponed until https://github.com/maplibre/maplibre-gl-js/issues/76 has been solved.
The performance benefit might be negligible, so having more concrete data would be required.

Just accepting this would unnecessarily complicate the codebase and imply that there's a benefit of doing this, when we are unable to confirm this right now.