@JannikGM 2021-11-22
I don't have time to look into this right now, as the company that I work for no longer prioritizes performance at the moment.
However, I'd be grateful if someone else could rebase and benchmark my changes.

I'll reopen this for exposure.