@HarelM 2021-11-22
@xabbu42 @wipfli Any chance you can take a look at this and see if it adds any benefit in terms of performance? I know you guys made the performance test possible...
I would say that this need a new PR instead of trying to fix this one as this was before the typescript change...