@wipfli 2021-11-22
I'm happy to mess around with this and learn something from Jannik.