@HarelM 2021-11-29
Closing this in favor of #666